Feat : 소켓 단일 페이지 → 전역 적용 수정
