+++ src/atoms/visitorAtroms.js
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const currentVisitorState = atom({
+  key: 'currentVisitorState',
+  default: 0,
+})
